requirements.txt
google-cloud-storage>=2.10.0
google-cloud-bigquery>=3.11.0
db-dtypes>=1.1.1
pandas>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
google-genai>=1.0.0
pyyaml>=6.0.0
google-cloud-secret-manager>=2.16.0
streamlit>=1.32.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
